requirements.txt
pytorch
transformers
ujson
tqdm